refactor: replace Breadcrumbs component with new implementation and update GrantsProposalReview for key assignment
