+++ src/components/BreadcrumbsNew/index.ts
@@ -1,0 +1,1 @@
+export { Breadcrumbs } from './Breadcrumbs'
