+++ src/components/MainContainer/index.ts
@@ -1,1 +1,2 @@
 export { MainContainer } from './MainContainer'
+export { menuData } from './sidebars/menuData'
